+++ p2p/providers/peerstore_provider/peerstore_provider.go
@@ -3,14 +3,14 @@ package peerstore_provider
 //go:generate mockgen -destination=../../types/mocks/peerstore_provider_mock.go -package=mock_types github.com/pokt-network/pocket/p2p/providers/peerstore_provider PeerstoreProvider
 
 import (
+	"go.uber.org/multierr"
+
 	"github.com/pokt-network/pocket/logger"
 	typesP2P "github.com/pokt-network/pocket/p2p/types"
 	"github.com/pokt-network/pocket/runtime/configs"
 	coreTypes "github.com/pokt-network/pocket/shared/core/types"
 	cryptoPocket "github.com/pokt-network/pocket/shared/crypto"
 	"github.com/pokt-network/pocket/shared/modules"
-	sharedP2P "github.com/pokt-network/pocket/shared/p2p"
-	"go.uber.org/multierr"
 )
 
 const ModuleName = "peerstore_provider"
@@ -19,14 +19,12 @@ const ModuleName = "peerstore_provider"
 type PeerstoreProvider interface {
 	modules.Module
 
-	GetStakedPeerstoreAtHeight(height uint64) (sharedP2P.Peerstore, error)
-	GetConnFactory() typesP2P.ConnectionFactory
+	GetStakedPeerstoreAtHeight(height uint64) (typesP2P.Peerstore, error)
 	GetP2PConfig() *configs.P2PConfig
-	SetConnectionFactory(typesP2P.ConnectionFactory)
 }
 
-func ActorsToPeerstore(abp PeerstoreProvider, actors []*coreTypes.Actor) (pstore sharedP2P.Peerstore, errs error) {
-	pstore = make(sharedP2P.PeerAddrMap)
+func ActorsToPeerstore(abp PeerstoreProvider, actors []*coreTypes.Actor) (pstore typesP2P.Peerstore, errs error) {
+	pstore = make(typesP2P.PeerAddrMap)
 	for _, a := range actors {
 		networkPeer, err := ActorToPeer(abp, a)
 		// TECHDEBT(#519): consider checking for behaviour instead of type. For reference: https://github.com/pokt-network/pocket/pull/611#discussion_r1147476057
@@ -45,33 +43,17 @@ func ActorsToPeerstore(abp PeerstoreProvider, actors []*coreTypes.Actor) (pstore
 	return pstore, errs
 }
 
-func ActorToPeer(abp PeerstoreProvider, actor *coreTypes.Actor) (sharedP2P.Peer, error) {
-	// TECHDEBT(#576): this should be the responsibility of some new `ConnManager` interface.
-	// Peerstore (identity / address mapping) is a separate concern from managing
-	// connections to/from peers.
-	conn, err := abp.GetConnFactory()(abp.GetP2PConfig(), actor.GetServiceUrl()) // generic param is service url
-	if err != nil {
-		return nil, NewErrResolvingAddr(err)
-	}
-
+func ActorToPeer(abp PeerstoreProvider, actor *coreTypes.Actor) (typesP2P.Peer, error) {
 	pubKey, err := cryptoPocket.NewPublicKey(actor.GetPublicKey())
 	if err != nil {
 		return nil, err
 	}
 
 	peer := &typesP2P.NetworkPeer{
-		Transport:  conn,
 		PublicKey:  pubKey,
 		Address:    pubKey.Address(),
 		ServiceURL: actor.GetServiceUrl(), // service url
 	}
 
 	return peer, nil
-}
-
-// WithConnectionFactory allows the user to specify a custom connection factory
-func WithConnectionFactory(connFactory typesP2P.ConnectionFactory) func(PeerstoreProvider) {
-	return func(ap PeerstoreProvider) {
-		ap.SetConnectionFactory(connFactory)
-	}
 }