+++ p2p/types/network.go
@@ -1,11 +1,10 @@
 package types
 
-//go:generate mockgen -source=$GOFILE -destination=./mocks/network_mock.go github.com/pokt-network/pocket/p2p/types Network
+//go:generate mockgen -source=$GOFILE -destination=./mocks/network_mock.go -package=mock_types github.com/pokt-network/pocket/p2p/types Network
 
 import (
 	cryptoPocket "github.com/pokt-network/pocket/shared/crypto"
 	"github.com/pokt-network/pocket/shared/modules"
-	sharedP2P "github.com/pokt-network/pocket/shared/p2p"
 )
 
 // TECHDEBT(olshansky): When we delete `stdnetwork` and only go with `raintree`, this interface
@@ -18,9 +17,9 @@ type Network interface {
 
 	// Address book helpers
 	// TECHDEBT: simplify - remove `GetPeerstore`
-	GetPeerstore() sharedP2P.Peerstore
-	AddPeer(peer sharedP2P.Peer) error
-	RemovePeer(peer sharedP2P.Peer) error
+	GetPeerstore() Peerstore
+	AddPeer(peer Peer) error
+	RemovePeer(peer Peer) error
 
 	// This function was added to specifically support the RainTree implementation.
 	// Handles the raw data received from the network and returns the data to be processed